Add support for Pin<P>
Add support for Pin pointer types to DefaultPointerOps so that users can
use a type like Pin<Box<T>> inside the intrusive_adapter! macro.
Intrusive collections inherently rely on the values having stable memory
locations and don't provide a way to get a mutable reference to the
value so they uphold the guarantees of pinning.